+++ src/operators/query/evaluation/index.ts
@@ -1,4 +1,5 @@
 export * from "./expr";
+export * from "./jsonSchema";
 export * from "./mod";
 export * from "./regex";
 // $where operator is not included by default